+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Django build does not include cyberscript shared library
 - Extend and streamline NGFW stand up plan
 - Dynamic subnet creation for ranges misses Shifter Platform creation
+- Missing VPC route for kali
 
 ## [0.10.7] - 2026-01-12